refactor: relocate shared DataFusion infrastructure to customscan/datafusion/
JoinScan and AggregateScan both reach into joinscan::{memory, translator,
explain} for code that has nothing to do with joins specifically — it's
the DataFusion glue both scans share. Forcing aggregatescan to depend on
joinscan as a sibling for that code is an abstraction smell that makes
the module hierarchy lie about the dependency graph.

Move the three whole-file consumers into a new neutral
`pg_search/src/postgres/customscan/datafusion/` namespace:

- joinscan/memory.rs    -> datafusion/memory.rs
- joinscan/translator.rs -> datafusion/translator.rs
- joinscan/explain.rs   -> datafusion/explain.rs

The move is a pure relocation — no logic changes. Intra-file references
that previously used `super::build` now reach into the unchanged
`joinscan::build` via absolute paths, since `RelNode`/`JoinSource`/
`JoinNode`/`RelationAlias`/`JoinLevelExpr`/`JoinCSClause` all still
live there for now.

`joinscan::privdat` is promoted from `mod` to `pub mod` so the moved
translator can still see `OutputColumnInfo` and `SCORE_COL_NAME`,
which `CombinedMapper` needs. Those types stay in joinscan because
they describe JoinScan-specific output column metadata; only the
visibility changes.

Consumers in joinscan (mod.rs, scan_state.rs, predicate.rs) and
aggregatescan (datafusion_exec.rs) have their imports rewritten to
the new paths.

Verify with `git diff --find-renames=80` — the file moves show as
pure renames.
