+++ pg_search/src/postgres/customscan/datafusion/memory.rs
